backend/embeddings/requirements.txt
absl-py==1.4.0
anyio==3.6.2
astunparse==1.6.3
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
fastapi==0.95.2
flatbuffers==23.5.9
gast==0.4.0
google-auth==2.18.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.2
h11==0.14.0
h5py==3.8.0
httptools==0.5.0
idna==3.4
keras==2.13.1rc0
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
minio==7.1.14
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
protobuf==4.23.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.7
python-dotenv==1.0.0
PyYAML==6.0
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tensorboard==2.13.0
tensorboard-data-server==0.7.0
tensorflow==2.13.0rc0
tensorflow-estimator==2.13.0rc0
tensorflow-hub==0.13.0
termcolor==2.3.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==2.3.4
wrapt==1.14.1
levenshtein==0.21.0
hiredis==2.2.3
redis==4.5.5